std::make_unique Pointer 관리 변경

+++ V1/Library/Dmi3dfInterface/3DF/Impl/Portfolio.Impl.cpp
@@ -131,7 +131,7 @@ bool H3DF::PortfolioKeyImpl::CreateShapeData(ShapeElementArray & arInShapeElemen
 			}
 
 			// Format code 추가
-			vfOutData.emplace_back(nFormatCode);
+			vfOutData.emplace_back((float)nFormatCode);
 
 			for (auto & cShapePoint : arShapePoints) {
 				CreateShapePointData(nFormatCode, cShapePoint, vfOutData);
